+++ docs/source/release-notes.rst
@@ -2,12 +2,17 @@
 Release Notes
 #############
 
-See what's been released in Vircadia. 
+See what's been released in Vircadia.
 
 .. toctree::
     :maxdepth: 2
     :titlesonly:
 
+    Release 2021.1.2 Eos (2021-05-30) <release-notes/2021-1-2-Eos>
+    Release 2021.1.1 Eos (2021-04-08) <release-notes/2021-1-1-Eos>
+    Release 2021.1.0 Eos (2021-02-21) <release-notes/2021-1-0-Eos>
+    Alpha Release 2020.3.3 Demeter (2021-01-02) <release-notes/2020-3-3-Demeter>
+    Alpha Release 2020.3.2 Demeter (2020-12-01) <release-notes/2020-3-2-Demeter>
     Alpha Release 2020.3.1 Demeter (2020-10-25) <release-notes/2020-3-1-Demeter>
     Alpha Release 2020.3.0 Demeter (2020-09-29) <release-notes/2020-3-0-Demeter>
     Alpha Release 2020.2.5 Asteria (2020-09-06) <release-notes/2020-2-5-Asteria>